+++ COT/commands/tests/test_command.py
@@ -16,10 +16,15 @@
 
 """Test cases for COT.commands.ReadWriteCommand class."""
 
+import os.path
+import mock
+
 from COT.tests.ut import COT_UT
 from COT.ui import UI
 from COT.commands import ReadWriteCommand
 from COT.vm_description import VMInitError
+
+# pylint: disable=missing-param-doc,missing-type-doc
 
 
 class TestReadWriteCommand(COT_UT):
@@ -43,8 +48,66 @@ class TestReadWriteCommand(COT_UT):
 
     def test_set_output_implicitly(self):
         """If 'output' is not specifically set, run() sets it to 'package'."""
-        self.instance = ReadWriteCommand(UI())
+        self.instance.output = ""
         self.instance.package = self.input_ovf
         self.assertEqual(self.instance.output, "")
         self.instance.run()
         self.assertEqual(self.instance.output, self.input_ovf)
+
+    @mock.patch("COT.vm_description.ovf.OVF.predicted_output_size")
+    @mock.patch("COT.commands.command.available_bytes_at_path")
+    def test_space_checks_called(self, mock_available, mock_size):
+        """Confirm that disk space checks are invoked appropriately."""
+        mock_size.return_value = 10000
+        mock_available.return_value = 50000
+
+        # Cases that do not result in a check:
+        # 1. Set package without setting output
+        self.instance = ReadWriteCommand(UI())
+        self.instance.package = self.input_ovf
+
+        mock_size.assert_not_called()
+        mock_available.assert_not_called()
+        self.instance.destroy()
+
+        # 2. Set output without setting package
+        self.instance = ReadWriteCommand(UI())
+        self.instance.output = self.temp_file
+        mock_size.assert_not_called()
+        mock_available.assert_not_called()
+
+        self.instance.destroy()
+        mock_size.reset_mock()
+        mock_available.reset_mock()
+
+        # Cases that do result in a check:
+        # 1. Setting package when output is already set
+        self.instance = ReadWriteCommand(UI())
+        self.instance.output = self.temp_file
+        self.instance.package = self.input_ovf
+        self.assertNotEqual(self.instance.vm, None)
+        mock_size.assert_called_once()
+        mock_available.assert_called_once()
+
+        self.instance.destroy()
+        mock_size.reset_mock()
+        mock_available.reset_mock()
+
+        # 2. Setting output when package is already set
+        self.instance = ReadWriteCommand(UI())
+        self.instance.package = self.input_ovf
+        self.instance.output = self.temp_file
+        self.assertNotEqual(self.instance.vm, None)
+        mock_size.assert_called_once()
+        mock_available.assert_called_once()
+
+        mock_size.reset_mock()
+        mock_available.reset_mock()
+
+        # 3. Changing output when package is already set
+        self.instance.output = os.path.join(self.temp_dir, "new_out.ovf")
+        mock_size.assert_called_once()
+        mock_available.assert_called_once()
+
+        mock_size.reset_mock()
+        mock_available.reset_mock()